Show dashboard on root for authenticated users
